Make runner check for explicit executable
The runner doesn't currently work for all languages, because how they differ in what exectable file they use.

This PR adds and argument to the run commands so that they point at the correct executable file.

I also change so that the input file is used for all benchmark. Even if it is an empty file for hello-world.

Plus, making the run.sh script take arguments that can override the the input file. To make it easier to test the runner with faster executing arguments.